Don't allow unfollow when the user doesn't already follow

+++ core/app/backbone/models/user.coffee
@@ -57,6 +57,8 @@ class window.User extends Backbone.Model
 
   unfollow: ->
     self = currentUser.following.get(@id)
+    return unless self
+
     self.destroy
       error: =>
         currentUser.following.add @
